add useEffect dependency and implement PriceField component
