+++ packages/web/app/utils/utils.jsx
@@ -103,17 +103,3 @@ export const preventInvalidNumber = event => {
     event.target.value = '';
   }
 };
-
-export const validateDecimalPlaces = (e, expectedDecimalPlaces = 2) => {
-  const value = e.target.value;
-  if (/^[−-]/.test(value)) {
-    e.target.value = '';
-    return;
-  }
-  if (value.includes('.')) {
-    const decimalPlaces = value.split('.')[1].length;
-    if (decimalPlaces > expectedDecimalPlaces) {
-      e.target.value = parseFloat(value).toFixed(expectedDecimalPlaces);
-    }
-  }
-};
